+++ components/headerSearchInput.tsx
@@ -1,14 +1,14 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
+import { useState } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
-import { useEffect, useState } from "react";
 
-const SearchInput = () => {
+const HeaderSearchInput = () => {
   const params = useSearchParams();
   const query = params.get("query") || "";
   const router = useRouter();
-  const [input, setInput] = useState(() => query);
+  const [input, setInput] = useState(query);
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setInput(e.target.value);
@@ -18,13 +18,6 @@ const SearchInput = () => {
     if (input.length < 1) return;
     router.push(`/reviews?query=${input}`);
   };
-  useEffect(() => {
-    // `params`가 바뀔 때마다 input 값을 갱신
-    const queryParam = params.get("query") || "";
-    if (queryParam !== input) {
-      setInput(queryParam); // URL에서 쿼리 값이 바뀌면 상태 업데이트
-    }
-  }, [params]); // params가 변경될 때마다 실행
 
   return (
     <div className="flex gap-2">
@@ -41,5 +34,4 @@ const SearchInput = () => {
     </div>
   );
 };
-
-export default SearchInput;
+export { HeaderSearchInput };
